// frontend/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        InterviewCreate: {
            track: components["schemas"]["InterviewTrackEnum"];
            level: components["schemas"]["InterviewLevelEnum"];
            /** @default technical */
            mode: components["schemas"]["InterviewModeEnum"];
            /**
             * Stress
             * @default false
             */
            stress: boolean;
            /**
             * Max Turns
             * @default 8
             */
            max_turns: number;
        };
// ...
        InterviewLevelEnum: "Junior" | "Middle" | "Senior";
// ...
        InterviewModeEnum: "technical" | "hr" | "mixed";
// ...
        InterviewTrackEnum: "Backend" | "Frontend" | "QA" | "DevOps";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}